% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean-Algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A number (int)}

\item{b}{A number (int)}
}
\value{
The greatest common divisor of both numbers.
}
\description{
Find the greatest common divisor of two numbers.
}
\details{
Pseudocode from: https://en.wikipedia.org/wiki/Euclidean_algorithm
}
\examples{
euclidean(100, 1000)
euclidean(123, 567890)
}
